Refactor prisma initiation
Postgres compute time was quickly being depleted during development.
I think this is because a new prisma client was being initiated on each refresh (code change).

This creates a singleton that is stored onto a global object, which can be reused between refreshes during development.

We'll see if it works 🤷🏻‍♂️